Guess number game - basic HTML , CSS , Js Project.
Embark on a thrilling journey of numbers with our "Guess the Number" game! This mini-project, crafted with HTML, JavaScript, and CSS, invites you to put your guessing skills to the test. The rules are simple: try to guess the correct number between 1 and 100, and if your guess aligns with the randomly generated number, congratulations – you've won the game!